Change group rule from string.quoted.regexp to string.regex.group
Not sure this is really the best choice, chose it because the Oniguruma syntax uses it. Seems to me that strings really shouldn't be inside themselves. Perhaps meta.? would be better for this?

Should really come up with a secondary set of conventions for regexp so we're on the same page everywhere.


git-svn-id: http://svn.textmate.org/trunk/Bundles/RegularExpressions.tmbundle@4184 dfb7d73b-c2ec-0310-8fea-fb051d288c6d